feat: add Ink account manager TUI
