+++ package.json
@@ -11,7 +11,7 @@
   },
   "scripts": {
     "build": "tsc -p tsconfig.build.json && node scripts/fix-esm-imports.mjs",
-    "format": "prettier --write 'src/**/*.ts'",
+    "format": "prettier --write 'src/**/*.{ts,tsx}'",
     "test": "bun test",
     "typecheck": "tsc --noEmit",
     "prepare": "husky"
@@ -38,14 +38,19 @@
   "dependencies": {
     "@aws/codewhisperer-streaming-client": "^1.0.34",
     "@opencode-ai/plugin": "^1.14.39",
+    "ink": "^7.0.2",
     "proper-lockfile": "^4.1.2",
+    "react": "^19.2.6",
+    "react-devtools-core": "^7.0.1",
     "zod": "^3.24.0"
   },
   "devDependencies": {
     "@types/node": "^20.0.0",
     "@types/proper-lockfile": "^4.1.4",
+    "@types/react": "^19.2.14",
     "bun-types": "^1.3.6",
     "husky": "^9.1.7",
+    "ink-testing-library": "^4.0.0",
     "lint-staged": "^16.2.7",
     "prettier": "^3.4.2",
     "prettier-plugin-organize-imports": "^4.1.0",
@@ -64,7 +69,7 @@
     "README.md"
   ],
   "lint-staged": {
-    "*.ts": [
+    "*.{ts,tsx}": [
       "prettier --write"
     ]
   }
